+++ runtime/src/bpf_map/array_map.cpp
@@ -1,0 +1,61 @@
+#include <bpf_map/array_map.hpp>
+namespace bpftime
+{
+array_map_impl::array_map_impl(
+	boost::interprocess::managed_shared_memory &memory, uint32_t value_size,
+	uint32_t max_entries)
+	: data(value_size * max_entries, memory.get_segment_manager())
+{
+	this->_value_size = value_size;
+	this->_max_entries = max_entries;
+}
+void *array_map_impl::elem_lookup(const void *key)
+{
+	auto key_val = *(uint32_t *)key;
+	if (key_val >= _max_entries) {
+		errno = ENOENT;
+		return nullptr;
+	}
+	return &data[key_val * _value_size];
+}
+
+long array_map_impl::elem_update(const void *key, const void *value,
+				 uint64_t flags)
+{
+	auto key_val = *(uint32_t *)key;
+	if (key_val >= _max_entries) {
+		errno = ENOENT;
+		return -1;
+	}
+	std::copy((uint8_t *)value, (uint8_t *)value + _value_size,
+		  &data[key_val * _value_size]);
+	return 0;
+}
+
+long array_map_impl::elem_delete(const void *key)
+{
+	auto key_val = *(uint32_t *)key;
+	if (key_val >= _max_entries) {
+		errno = ENOENT;
+		return -1;
+	}
+	std::fill(&data[key_val * _value_size],
+		  &data[key_val * _value_size] + _value_size, 0);
+	return 0;
+}
+
+int array_map_impl::bpf_map_get_next_key(const void *key, void *next_key)
+{
+	if (_max_entries == 0 || *(uint32_t *)key == _max_entries - 1) {
+		errno = ENOENT;
+		return -1;
+	}
+	if (key == nullptr || *(uint32_t *)key >= _max_entries - 1) {
+		*(uint32_t *)next_key = 0;
+		return 0;
+	}
+	auto key_val = *(uint32_t *)key;
+	*(uint32_t *)next_key = key_val + 1;
+	return 0;
+}
+} // namespace bpftime
